src/main/java/com/cesi/ressourcesrelationnelles/domain/Resource.java
package com.cesi.ressourcesrelationnelles.domain;

import java.util.Date;

public class Resource {
    private int id;
    private String title;
    private Date datePublication;
    private String url;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public Date getDatePublication() {
        return datePublication;
    }

    public void setDatePublication(Date datePublication) {
        this.datePublication = datePublication;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }
}
